+++ cryptic/extensions.py
@@ -2,7 +2,6 @@
 __author__ = 'fansly'
 
 from flask_bootstrap import Bootstrap
-from flask_ckeditor import CKEditor
 from flask_login import LoginManager
 from flask_mail import Mail
 from flask_moment import Moment
@@ -11,17 +10,19 @@ from flask_wtf import CSRFProtect
 from flask_debugtoolbar import DebugToolbarExtension
 from flask_migrate import Migrate
 from flask_whooshee import Whooshee
+from flask_qiniustorage import Qiniu
 
 bootstrap = Bootstrap()
 db = SQLAlchemy()
 login_manager = LoginManager()
 csrf = CSRFProtect()
-ckeditor = CKEditor()
 mail = Mail()
 moment = Moment()
 toolbar = DebugToolbarExtension()
 migrate = Migrate()
 whooshee = Whooshee()
+qiniu_store = Qiniu()
+
 
 @login_manager.user_loader
 def load_user(user_id):