+++ CHANGELOG.md
@@ -6,11 +6,45 @@ Demo: https://fancysly.com
 
 ## [Unreleased]
 ### Changed
-- Title Color change back to write
-- humans.txt
+- None
 
 ### Known Bug
 - the login model without entrance
+- without human.txt
+
+## [0.21] - 2018-11-22
+### Added
+- Add picture support
+- Add qiniu upload support
+- mysql database support
+- support Alembic
+
+### Changed
+- the editor.md file
+- remind not to login
+- the administrator manage entrance
+- the database model
+
+### Removed
+- elder markdown render(replaced by editor.md)
+- sqlite3 (replaced by mysql)
+- useless static folder
+
+
+## [0.1.3] - 2018-11-17
+### Added
+- Title Color change back to write
+- editor.md
+
+### Changed
+- Changed CKEditor4 to Markdown Editor
+- the Admin tansform bug
+- the text always in center
+
+### Removed
+- the icon to like me in github
+- the flask-Ckeditor
+
 
 ## [0.1.2] - 2018-11-15
 ### Added